@@ -1,16 +1,35 @@
+//==============================================================================
+// This file is part of Master Password.
+// Copyright (c) 2011-2017, Maarten Billemont.
+//
+// Master Password is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Master Password is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You can find a copy of the GNU General Public License in the
+// LICENSE file.  Alternatively, see <http://www.gnu.org/licenses/>.
+//==============================================================================
+
 package com.lyndir.masterpassword.gui.view;
 
-import static com.lyndir.lhunath.opal.system.util.StringUtils.strf;
+import static com.lyndir.lhunath.opal.system.util.StringUtils.*;
 
-import com.google.common.base.Function;
 import com.google.common.base.Preconditions;
-import com.google.common.collect.*;
+import com.google.common.collect.ImmutableList;
+import com.google.common.primitives.UnsignedInteger;
 import com.lyndir.lhunath.opal.system.logging.Logger;
+import com.lyndir.masterpassword.MPAlgorithm;
+import com.lyndir.masterpassword.MPResultType;
 import com.lyndir.masterpassword.gui.Res;
-import com.lyndir.masterpassword.gui.model.ModelUser;
-import com.lyndir.masterpassword.model.MPUser;
-import com.lyndir.masterpassword.model.MPUserFileManager;
 import com.lyndir.masterpassword.gui.util.Components;
+import com.lyndir.masterpassword.model.MPUser;
+import com.lyndir.masterpassword.model.impl.*;
 import java.awt.*;
 import java.awt.event.*;
 import javax.annotation.Nonnull;
@@ -24,14 +43,15 @@ import javax.swing.plaf.metal.MetalComboBoxEditor;
 /**
  * @author lhunath, 2014-06-11
  */
-public class ModelAuthenticationPanel extends AuthenticationPanel implements ItemListener, ActionListener, DocumentListener {
+public class ModelAuthenticationPanel extends AuthenticationPanel<MPFileUser> implements ItemListener, ActionListener, DocumentListener {
 
     @SuppressWarnings("UnusedDeclaration")
-    private static final Logger logger = Logger.get( ModelAuthenticationPanel.class );
+    private static final Logger logger           = Logger.get( ModelAuthenticationPanel.class );
+    private static final long   serialVersionUID = 1L;
 
-    private final JComboBox<ModelUser> userField;
-    private final JLabel               masterPasswordLabel;
-    private final JPasswordField       masterPasswordField;
+    private final JComboBox<MPFileUser> userField;
+    private final JLabel                masterPasswordLabel;
+    private final JPasswordField        masterPasswordField;
 
     public ModelAuthenticationPanel(final UnlockFrame unlockFrame) {
         super( unlockFrame );
@@ -41,7 +61,7 @@ public class ModelAuthenticationPanel extends AuthenticationPanel implements Ite
         avatarLabel.addMouseListener( new MouseAdapter() {
             @Override
             public void mouseClicked(final MouseEvent e) {
-                ModelUser selectedUser = getSelectedUser();
+                MPFileUser selectedUser = getSelectedUser();
                 if (selectedUser != null) {
                     selectedUser.setAvatar( selectedUser.getAvatar() + 1 );
                     updateUser( false );
@@ -54,9 +74,20 @@ public class ModelAuthenticationPanel extends AuthenticationPanel implements Ite
         add( userLabel );
 
         userField = Components.comboBox( readConfigUsers() );
-        userField.setFont( Res.valueFont().deriveFont( 12f ) );
+        userField.setFont( Res.valueFont().deriveFont( userField.getFont().getSize2D() ) );
         userField.addItemListener( this );
         userField.addActionListener( this );
+        userField.setRenderer( new DefaultListCellRenderer() {
+            private static final long serialVersionUID = 1L;
+
+            @Override
+            @SuppressWarnings("unchecked")
+            public Component getListCellRendererComponent(final JList<?> list, final Object value, final int index,
+                                                          final boolean isSelected, final boolean cellHasFocus) {
+                String userValue = ((MPUser<MPFileSite>) value).getFullName();
+                return super.getListCellRendererComponent( list, userValue, index, isSelected, cellHasFocus );
+            }
+        } );
         userField.setEditor( new MetalComboBoxEditor() {
             @Override
             protected JTextField createEditorComponent() {
@@ -86,10 +117,10 @@ public class ModelAuthenticationPanel extends AuthenticationPanel implements Ite
 
     @Override
     protected void updateUser(boolean repack) {
-        ModelUser selectedUser = getSelectedUser();
+        MPFileUser selectedUser = getSelectedUser();
         if (selectedUser != null) {
             avatarLabel.setIcon( Res.avatar( selectedUser.getAvatar() ) );
-            boolean showPasswordField = !selectedUser.keySaved();
+            boolean showPasswordField = !selectedUser.isMasterKeyAvailable(); // TODO: is this the same as keySaved()?
             if (masterPasswordField.isVisible() != showPasswordField) {
                 masterPasswordLabel.setVisible( showPasswordField );
                 masterPasswordField.setVisible( showPasswordField );
@@ -100,8 +131,9 @@ public class ModelAuthenticationPanel extends AuthenticationPanel implements Ite
         super.updateUser( repack );
     }
 
+    @Nullable
     @Override
-    protected ModelUser getSelectedUser() {
+    protected MPFileUser getSelectedUser() {
         int selectedIndex = userField.getSelectedIndex();
         if (selectedIndex < 0)
             return null;
@@ -125,7 +157,7 @@ public class ModelAuthenticationPanel extends AuthenticationPanel implements Ite
                         String fullName = JOptionPane.showInputDialog( ModelAuthenticationPanel.this, //
                                                                        "Enter your full name, ensuring it is correctly spelled and capitalized:",
                                                                        "New User", JOptionPane.QUESTION_MESSAGE );
-                        MPUserFileManager.get().addUser( new MPUser( fullName ) );
+                        MPFileUserManager.get().addUser( new MPFileUser( fullName ) );
                         userField.setModel( new DefaultComboBoxModel<>( readConfigUsers() ) );
                         updateUser( true );
                     }
@@ -137,17 +169,18 @@ public class ModelAuthenticationPanel extends AuthenticationPanel implements Ite
                 addActionListener( new ActionListener() {
                     @Override
                     public void actionPerformed(final ActionEvent e) {
-                        ModelUser deleteUser = getSelectedUser();
+                        MPFileUser deleteUser = getSelectedUser();
                         if (deleteUser == null)
                             return;
 
                         if (JOptionPane.showConfirmDialog( ModelAuthenticationPanel.this, //
-                                                       strf( "Are you sure you want to delete the user and sites remembered for:\n%s.",
-                                                             deleteUser.getFullName() ), //
-                                                       "Delete User", JOptionPane.OK_CANCEL_OPTION, JOptionPane.QUESTION_MESSAGE ) == JOptionPane.CANCEL_OPTION)
+                                                           strf( "Are you sure you want to delete the user and sites remembered for:%n%s.",
+                                                                 deleteUser.getFullName() ), //
+                                                           "Delete User", JOptionPane.OK_CANCEL_OPTION, JOptionPane.QUESTION_MESSAGE )
+                            == JOptionPane.CANCEL_OPTION)
                             return;
 
-                        MPUserFileManager.get().deleteUser( deleteUser.getModel() );
+                        MPFileUserManager.get().deleteUser( deleteUser );
                         userField.setModel( new DefaultComboBoxModel<>( readConfigUsers() ) );
                         updateUser( true );
                     }
@@ -156,15 +189,11 @@ public class ModelAuthenticationPanel extends AuthenticationPanel implements Ite
             }
         }, new JButton( Res.iconQuestion() ) {
             {
-                addActionListener( new ActionListener() {
-                    @Override
-                    public void actionPerformed(final ActionEvent e) {
-                        JOptionPane.showMessageDialog( ModelAuthenticationPanel.this, //
-                                                       strf( "Reads users and sites from the directory at:\n%s",
-                                                             MPUserFileManager.get().getPath().getAbsolutePath() ), //
-                                                       "Help", JOptionPane.INFORMATION_MESSAGE );
-                    }
-                } );
+                addActionListener( e -> JOptionPane.showMessageDialog(
+                        ModelAuthenticationPanel.this, //
+                        strf( "Reads users and sites from the directory at:%n%s",
+                              MPFileUserManager.get().getPath().getAbsolutePath() ), //
+                        "Help", JOptionPane.INFORMATION_MESSAGE ) );
                 setToolTipText( "More information." );
             }
         } );
@@ -175,14 +204,20 @@ public class ModelAuthenticationPanel extends AuthenticationPanel implements Ite
         masterPasswordField.setText( "" );
     }
 
-    private static ModelUser[] readConfigUsers() {
-        return FluentIterable.from( MPUserFileManager.get().getUsers() ).transform( new Function<MPUser, ModelUser>() {
-            @Nullable
+    @Override
+    public PasswordFrame<MPFileUser, MPFileSite> newPasswordFrame() {
+        return new PasswordFrame<MPFileUser, MPFileSite>( Preconditions.checkNotNull( getSelectedUser() ) ) {
             @Override
-            public ModelUser apply(@Nullable final MPUser model) {
-                return new ModelUser( Preconditions.checkNotNull( model ) );
+            protected MPFileSite createSite(final MPFileUser user, final String siteName, final UnsignedInteger siteCounter,
+                                            final MPResultType resultType,
+                                            final MPAlgorithm algorithm) {
+                return new MPFileSite( user, siteName, algorithm, siteCounter, resultType );
             }
-        } ).toArray( ModelUser.class );
+        };
+    }
+
+    private static MPFileUser[] readConfigUsers() {
+        return MPFileUserManager.get().getUsers().toArray( new MPFileUser[0] );
     }
 
     @Override
