@@ -1,12 +1,31 @@
+//==============================================================================
+// This file is part of Master Password.
+// Copyright (c) 2011-2017, Maarten Billemont.
+//
+// Master Password is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Master Password is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You can find a copy of the GNU General Public License in the
+// LICENSE file.  Alternatively, see <http://www.gnu.org/licenses/>.
+//==============================================================================
+
 package com.lyndir.masterpassword.gui.view;
 
 import static com.lyndir.lhunath.opal.system.util.ObjectUtils.*;
+import static com.lyndir.lhunath.opal.system.util.StringUtils.*;
 
+import com.lyndir.masterpassword.MPAlgorithmException;
 import com.lyndir.masterpassword.MPIdenticon;
-import com.lyndir.masterpassword.gui.*;
-import com.lyndir.masterpassword.gui.model.User;
+import com.lyndir.masterpassword.gui.Res;
 import com.lyndir.masterpassword.gui.util.Components;
-import com.lyndir.masterpassword.model.IncorrectMasterPasswordException;
+import com.lyndir.masterpassword.model.*;
 import java.awt.*;
 import java.awt.event.*;
 import java.util.concurrent.Future;
@@ -18,17 +37,19 @@ import javax.swing.*;
 /**
  * @author lhunath, 2014-06-08
  */
+@SuppressWarnings({ "MagicNumber", "serial" })
 public class UnlockFrame extends JFrame {
 
-    private final SignInCallback           signInCallback;
-    private final Components.GradientPanel root;
-    private final JLabel                   identiconLabel;
-    private final JButton                  signInButton;
-    private final JPanel                   authenticationContainer;
-    private       AuthenticationPanel      authenticationPanel;
-    private       Future<?>                identiconFuture;
-    private       boolean                  incognito;
-    private       User                     user;
+    private final SignInCallback              signInCallback;
+    private final Components.GradientPanel    root;
+    private final JLabel                      identiconLabel;
+    private final JButton                     signInButton;
+    private final JPanel                      authenticationContainer;
+    private       AuthenticationPanel<?>      authenticationPanel;
+    private       Future<?>                   identiconFuture;
+    private       boolean                     incognito;
+    @Nullable
+    private       MPUser<? extends MPSite<?>> user;
 
     public UnlockFrame(final SignInCallback signInCallback) {
         super( "Unlock Master Password" );
@@ -67,7 +88,7 @@ public class UnlockFrame extends JFrame {
         authenticationContainer.setBorder( BorderFactory.createEmptyBorder( 20, 20, 20, 20 ) );
         identiconLabel.setFont( Res.emoticonsFont().deriveFont( 14.f ) );
         identiconLabel.setToolTipText(
-                "A representation of your identity across all Master Password apps.\nIt should always be the same." );
+                strf( "A representation of your identity across all Master Password apps.%nIt should always be the same." ) );
         signInButton.addActionListener( new AbstractAction() {
             @Override
             public void actionPerformed(final ActionEvent e) {
@@ -99,20 +120,12 @@ public class UnlockFrame extends JFrame {
         authenticationContainer.add( authenticationPanel );
         authenticationContainer.add( Components.stud() );
 
-        final JCheckBox incognitoCheckBox = Components.checkBox( "Incognito" );
+        JCheckBox incognitoCheckBox = Components.checkBox( "Incognito" );
         incognitoCheckBox.setToolTipText( "Log in without saving any information." );
         incognitoCheckBox.setSelected( incognito );
-        incognitoCheckBox.addItemListener( new ItemListener() {
-            @Override
-            public void itemStateChanged(final ItemEvent e) {
-                incognito = incognitoCheckBox.isSelected();
-                SwingUtilities.invokeLater( new Runnable() {
-                    @Override
-                    public void run() {
-                        createAuthenticationPanel();
-                    }
-                } );
-            }
+        incognitoCheckBox.addItemListener( e -> {
+            incognito = incognitoCheckBox.isSelected();
+            SwingUtilities.invokeLater( this::createAuthenticationPanel );
         } );
 
         JComponent toolsPanel = Components.boxLayout( BoxLayout.LINE_AXIS, incognitoCheckBox, Box.createGlue() );
@@ -129,15 +142,10 @@ public class UnlockFrame extends JFrame {
         validate();
         repack();
 
-        SwingUtilities.invokeLater( new Runnable() {
-            @Override
-            public void run() {
-                ifNotNullElse( authenticationPanel.getFocusComponent(), signInButton ).requestFocusInWindow();
-            }
-        } );
+        SwingUtilities.invokeLater( () -> ifNotNullElse( authenticationPanel.getFocusComponent(), signInButton ).requestFocusInWindow() );
     }
 
-    void updateUser(@Nullable final User user) {
+    void updateUser(@Nullable final MPUser<? extends MPSite<?>> user) {
         this.user = user;
         checkSignIn();
     }
@@ -145,39 +153,31 @@ public class UnlockFrame extends JFrame {
     boolean checkSignIn() {
         if (identiconFuture != null)
             identiconFuture.cancel( false );
-        identiconFuture = Res.schedule( this, new Runnable() {
-            @Override
-            public void run() {
-                SwingUtilities.invokeLater( new Runnable() {
-                    @Override
-                    public void run() {
-                        String fullName = (user == null)? "": user.getFullName();
-                        char[] masterPassword = authenticationPanel.getMasterPassword();
+        identiconFuture = Res.schedule( this, () -> SwingUtilities.invokeLater( () -> {
+            String fullName       = (user == null)? "": user.getFullName();
+            char[] masterPassword = authenticationPanel.getMasterPassword();
 
-                        if (fullName.isEmpty() || (masterPassword.length == 0)) {
-                            identiconLabel.setText( " " );
-                            return;
-                        }
-
-                        MPIdenticon identicon = new MPIdenticon( fullName, masterPassword );
-                        identiconLabel.setText( identicon.getText() );
-                        identiconLabel.setForeground(
-                                Res.colors().fromIdenticonColor( identicon.getColor(), Res.Colors.BackgroundMode.DARK ) );
-                    }
-                } );
+            if (fullName.isEmpty() || (masterPassword.length == 0)) {
+                identiconLabel.setText( " " );
+                return;
             }
-        }, 300, TimeUnit.MILLISECONDS );
 
-        String fullName = (user == null)? "": user.getFullName();
-        char[] masterPassword = authenticationPanel.getMasterPassword();
-        boolean enabled = !fullName.isEmpty() && (masterPassword.length > 0);
+            MPIdenticon identicon = new MPIdenticon( fullName, masterPassword );
+            identiconLabel.setText( identicon.getText() );
+            identiconLabel.setForeground(
+                    Res.colors().fromIdenticonColor( identicon.getColor(), Res.Colors.BackgroundMode.DARK ) );
+        } ), 300, TimeUnit.MILLISECONDS );
+
+        String  fullName       = (user == null)? "": user.getFullName();
+        char[]  masterPassword = authenticationPanel.getMasterPassword();
+        boolean enabled        = !fullName.isEmpty() && (masterPassword.length > 0);
         signInButton.setEnabled( enabled );
 
         return enabled;
     }
 
     void trySignIn(final JComponent... signInComponents) {
-        if (!checkSignIn())
+        if ((user == null) || !checkSignIn())
             return;
 
         for (final JComponent signInComponent : signInComponents)
@@ -186,39 +186,31 @@ public class UnlockFrame extends JFrame {
         signInButton.setEnabled( false );
         signInButton.setText( "Signing In..." );
 
-        Res.execute( this, new Runnable() {
-            @Override
-            public void run() {
-                try {
-                    user.authenticate( authenticationPanel.getMasterPassword() );
+        Res.execute( this, () -> {
+            try {
+                user.authenticate( authenticationPanel.getMasterPassword() );
 
-                    SwingUtilities.invokeLater( new Runnable() {
-                        @Override
-                        public void run() {
-                            signInCallback.signedIn( user );
-                            dispose();
-                        }
-                    } );
-                }
-                catch (final IncorrectMasterPasswordException e) {
-                    SwingUtilities.invokeLater( new Runnable() {
-                        @Override
-                        public void run() {
-                            JOptionPane.showMessageDialog( null, e.getLocalizedMessage(), "Sign In Failed", JOptionPane.ERROR_MESSAGE );
-                            authenticationPanel.reset();
-                            signInButton.setText( "Sign In" );
-                            for (final JComponent signInComponent : signInComponents)
-                                signInComponent.setEnabled( true );
-                            checkSignIn();
-                        }
-                    } );
-                }
+                SwingUtilities.invokeLater( () -> {
+                    signInCallback.signedIn( authenticationPanel.newPasswordFrame() );
+                    dispose();
+                } );
+            }
+            catch (final MPIncorrectMasterPasswordException | MPAlgorithmException e) {
+                SwingUtilities.invokeLater( () -> {
+                    JOptionPane.showMessageDialog( null, e.getLocalizedMessage(), "Sign In Failed", JOptionPane.ERROR_MESSAGE );
+                    authenticationPanel.reset();
+                    signInButton.setText( "Sign In" );
+                    for (final JComponent signInComponent : signInComponents)
+                        signInComponent.setEnabled( true );
+                    checkSignIn();
+                } );
             }
         } );
     }
 
+    @FunctionalInterface
     public interface SignInCallback {
 
-        void signedIn(User user);
+        void signedIn(PasswordFrame<?, ?> passwordFrame);
     }
 }
