@@ -17,11 +17,10 @@
 //==============================================================================
 
 /**
- *
  * @author lhunath, 15-02-04
  */
 
-
-@ParametersAreNonnullByDefault package com.lyndir.masterpassword;
+@ParametersAreNonnullByDefault
+package com.lyndir.masterpassword;
 
 import javax.annotation.ParametersAreNonnullByDefault;
