@@ -22,8 +22,10 @@ import android.os.Handler;
 import android.os.Looper;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Sets;
-import java.util.*;
+import java.util.List;
+import java.util.Set;
 import java.util.concurrent.*;
+import javax.annotation.Nonnull;
 
 
 /**
@@ -33,7 +35,7 @@ public class MainThreadExecutor extends AbstractExecutorService {
 
     private final Handler       mHandler = new Handler( Looper.getMainLooper() );
     private final Set<Runnable> commands = Sets.newLinkedHashSet();
-    private boolean shutdown;
+    private       boolean       shutdown;
 
     @Override
     public void execute(final Runnable command) {
@@ -63,6 +65,7 @@ public class MainThreadExecutor extends AbstractExecutorService {
         shutdown = true;
     }
 
+    @Nonnull
     @Override
     public List<Runnable> shutdownNow() {
         shutdown = true;
