@@ -225,7 +225,7 @@
 - (void)copyAnswer:(NSString *)answer {
 
     UIPasteboard *pasteboard = [UIPasteboard generalPasteboard];
-    if ([pasteboard respondsToSelector:@selector( setItems:options: )]) {
+    if (@available(iOS 10.0, *)) {
         [pasteboard setItems:@[ @{ UIPasteboardTypeAutomatic: answer } ]
                      options:@{
                              UIPasteboardOptionLocalOnly     : @NO,
