@@ -16,6 +16,8 @@
 // LICENSE file.  Alternatively, see <http://www.gnu.org/licenses/>.
 //==============================================================================
 
+#define _POSIX_C_SOURCE 200809L
+
 #include <stdio.h>
 #include <unistd.h>
 #include <stdlib.h>
@@ -26,32 +28,32 @@
 #include "mpw-cli-util.h"
 #include "mpw-algorithm.h"
 #include "mpw-util.h"
-#include "mpw-marshall.h"
+#include "mpw-marshal.h"
 
 /** Output the program's usage documentation. */
 static void usage() {
 
     inf( ""
-            "  Master Password v%s\n"
+            "  Master Password v%s - CLI\n"
             "--------------------------------------------------------------------------------\n"
-            "      https://masterpasswordapp.com\n\n", stringify_def( MP_VERSION ) );
+            "      https://masterpassword.app\n", stringify_def( MP_VERSION ) );
     inf( ""
             "\nUSAGE\n\n"
             "  mpw [-u|-U full-name] [-m fd] [-t pw-type] [-P value] [-c counter]\n"
             "      [-a version] [-p purpose] [-C context] [-f|F format] [-R 0|1]\n"
-            "      [-v|-q] [-h] [site-name]\n\n" );
+            "      [-v|-q]* [-h] [site-name]\n" );
     inf( ""
             "  -u full-name Specify the full name of the user.\n"
             "               -u checks the master password against the config,\n"
             "               -U allows updating to a new master password.\n"
-            "               Defaults to %s in env or prompts.\n\n", MP_ENV_fullName );
+            "               Defaults to %s in env or prompts.\n", MP_ENV_fullName );
     dbg( ""
             "  -M master-pw Specify the master password of the user.\n"
-            "               Passing secrets as arguments is unsafe, for use in testing only.\n" );
+            "               Passing secrets as arguments is unsafe, for use in testing only." );
     inf( ""
             "  -m fd        Read the master password of the user from a file descriptor.\n"
             "               Tip: don't send extra characters like newlines such as by using\n"
-            "               echo in a pipe.  Consider printf instead.\n\n" );
+            "               echo in a pipe.  Consider printf instead.\n" );
     inf( ""
             "  -t pw-type   Specify the password's template.\n"
             "               Defaults to 'long' (-p a), 'name' (-p i) or 'phrase' (-p r).\n"
@@ -63,57 +65,59 @@ static void usage() {
             "                   i, pin      | 4 numbers.\n"
             "                   n, name     | 9 letter name.\n"
             "                   p, phrase   | 20 character sentence.\n"
-            "                   K, key      | encryption key (set key size -s bits).\n"
-            "                   P, personal | saved personal password (save with -s pw).\n\n" );
+            "                   K, key      | encryption key (512 bit or -P bits).\n"
+            "                   P, personal | saved personal password (save with -P pw).\n" );
     inf( ""
             "  -P value     The parameter value.\n"
             "                   -p i        | The login name for the site.\n"
             "                   -t K        | The bit size of the key to generate (eg. 256).\n"
-            "                   -t P        | The personal password to encrypt.\n\n" );
+            "                   -t P        | The personal password to encrypt.\n" );
     inf( ""
             "  -c counter   The value of the counter.\n"
-            "               Defaults to 1.\n\n" );
+            "               Defaults to 1.\n" );
     inf( ""
             "  -a version   The algorithm version to use, %d - %d.\n"
-            "               Defaults to %s in env or %d.\n\n",
+            "               Defaults to env var %s or %d.\n",
             MPAlgorithmVersionFirst, MPAlgorithmVersionLast, MP_ENV_algorithm, MPAlgorithmVersionCurrent );
     inf( ""
             "  -p purpose   The purpose of the generated token.\n"
             "               Defaults to 'auth'.\n"
             "                   a, auth     | An authentication token such as a password.\n"
             "                   i, ident    | An identification token such as a username.\n"
-            "                   r, rec      | A recovery token such as a security answer.\n\n" );
+            "                   r, rec      | A recovery token such as a security answer.\n" );
     inf( ""
             "  -C context   A purpose-specific context.\n"
             "               Defaults to empty.\n"
             "                   -p a        | -\n"
             "                   -p i        | -\n"
-            "                   -p r        | Most significant word in security question.\n\n" );
+            "                   -p r        | Most significant word in security question.\n" );
     inf( ""
             "  -f|F format  The mpsites format to use for reading/writing site parameters.\n"
             "               -F forces the use of the given format,\n"
             "               -f allows fallback/migration.\n"
-            "               Defaults to %s in env or json, falls back to plain.\n"
+            "               Defaults to env var %s or json, falls back to plain.\n"
             "                   n, none     | No file\n"
             "                   f, flat     | ~/.mpw.d/Full Name.%s\n"
-            "                   j, json     | ~/.mpw.d/Full Name.%s\n\n",
-            MP_ENV_format, mpw_marshall_format_extension( MPMarshallFormatFlat ), mpw_marshall_format_extension( MPMarshallFormatJSON ) );
+            "                   j, json     | ~/.mpw.d/Full Name.%s\n",
+            MP_ENV_format, mpw_marshal_format_extension( MPMarshalFormatFlat ), mpw_marshal_format_extension( MPMarshalFormatJSON ) );
     inf( ""
             "  -R redacted  Whether to save the mpsites in redacted format or not.\n"
             "               Redaction omits or encrypts any secrets, making the file safe\n"
             "               for saving on or transmitting via untrusted media.\n"
-            "               Defaults to 1, redacted.\n\n" );
+            "               Defaults to 1, redacted.\n" );
     inf( ""
             "  -v           Increase output verbosity (can be repeated).\n"
-            "  -q           Decrease output verbosity (can be repeated).\n\n" );
+            "  -q           Decrease output verbosity (can be repeated).\n" );
     inf( ""
-            "  -h           Show this help output instead of performing any operation.\n\n" );
+            "  -h           Show this help output instead of performing any operation.\n" );
+    inf( ""
+            "  site-name    Name of the site for which to generate a token.\n" );
     inf( ""
             "\nENVIRONMENT\n\n"
             "  %-12s The full name of the user (see -u).\n"
             "  %-12s The default algorithm version (see -a).\n"
             "  %-12s The default mpsites format (see -f).\n"
-            "  %-12s The askpass program to use for prompting the user.\n\n",
+            "  %-12s The askpass program to use for prompting the user.\n",
             MP_ENV_fullName, MP_ENV_algorithm, MP_ENV_format, MP_ENV_askpass );
     exit( EX_OK );
 }
@@ -142,7 +146,7 @@ typedef struct {
     const char *masterPassword;
     const char *identicon;
     const char *siteName;
-    MPMarshallFormat sitesFormat;
+    MPMarshalFormat sitesFormat;
     MPKeyPurpose keyPurpose;
     const char *keyContext;
     const char *sitesPath;
@@ -191,7 +195,7 @@ int main(const int argc, char *const argv[]) {
     Operation operation = {
             .allowPasswordUpdate = false,
             .sitesFormatFixed = false,
-            .sitesFormat = MPMarshallFormatDefault,
+            .sitesFormat = MPMarshalFormatDefault,
             .keyPurpose = MPKeyPurposeAuthentication,
             .resultType = MPResultTypeDefault,
             .siteCounter = MPCounterValueDefault,
@@ -223,24 +227,24 @@ int main(const int argc, char *const argv[]) {
     cli_free( &args, NULL );
 
     // Operation summary.
-    dbg( "-----------------\n" );
+    dbg( "-----------------" );
     if (operation.user) {
-        dbg( "fullName         : %s\n", operation.user->fullName );
-        trc( "masterPassword   : %s\n", operation.user->masterPassword );
-        dbg( "identicon        : %s\n", operation.identicon );
-        dbg( "sitesFormat      : %s%s\n", mpw_nameForFormat( operation.sitesFormat ), operation.sitesFormatFixed? " (fixed)": "" );
-        dbg( "sitesPath        : %s\n", operation.sitesPath );
+        dbg( "fullName         : %s", operation.user->fullName );
+        trc( "masterPassword   : %s", operation.user->masterPassword );
+        dbg( "identicon        : %s", operation.identicon );
+        dbg( "sitesFormat      : %s%s", mpw_nameForFormat( operation.sitesFormat ), operation.sitesFormatFixed? " (fixed)": "" );
+        dbg( "sitesPath        : %s", operation.sitesPath );
     }
     if (operation.site) {
-        dbg( "siteName         : %s\n", operation.site->name );
-        dbg( "siteCounter      : %u\n", operation.siteCounter );
-        dbg( "resultType       : %s (%u)\n", mpw_nameForType( operation.resultType ), operation.resultType );
-        dbg( "resultParam      : %s\n", operation.resultParam );
-        dbg( "keyPurpose       : %s (%u)\n", mpw_nameForPurpose( operation.keyPurpose ), operation.keyPurpose );
-        dbg( "keyContext       : %s\n", operation.keyContext );
-        dbg( "algorithmVersion : %u\n", operation.site->algorithm );
+        dbg( "siteName         : %s", operation.site->name );
+        dbg( "siteCounter      : %u", operation.siteCounter );
+        dbg( "resultType       : %s (%u)", mpw_nameForType( operation.resultType ), operation.resultType );
+        dbg( "resultParam      : %s", operation.resultParam );
+        dbg( "keyPurpose       : %s (%u)", mpw_nameForPurpose( operation.keyPurpose ), operation.keyPurpose );
+        dbg( "keyContext       : %s", operation.keyContext );
+        dbg( "algorithmVersion : %u", operation.site->algorithm );
     }
-    dbg( "-----------------\n\n" );
+    dbg( "-----------------" );
 
     // Finally ready to perform the actual operation.
     cli_mpw( &args, &operation );
@@ -272,51 +276,52 @@ void cli_free(Arguments *args, Operation *operation) {
 
 void cli_args(Arguments *args, Operation *operation, const int argc, char *const argv[]) {
 
-    for (int opt; (opt = getopt( argc, argv, "u:U:m:M:t:P:c:a:p:C:f:F:R:vqh" )) != EOF; optarg? bzero( optarg, strlen( optarg ) ): NULL)
+    for (int opt; (opt = getopt( argc, argv, "u:U:m:M:t:P:c:a:p:C:f:F:R:vqh" )) != EOF;
+         optarg? mpw_zero( optarg, strlen( optarg ) ): (void)0)
         switch (opt) {
             case 'u':
-                args->fullName = optarg && strlen( optarg )? strdup( optarg ): NULL;
+                args->fullName = optarg && strlen( optarg )? mpw_strdup( optarg ): NULL;
                 operation->allowPasswordUpdate = false;
                 break;
             case 'U':
-                args->fullName = optarg && strlen( optarg )? strdup( optarg ): NULL;
+                args->fullName = optarg && strlen( optarg )? mpw_strdup( optarg ): NULL;
                 operation->allowPasswordUpdate = true;
                 break;
             case 'm':
-                args->masterPasswordFD = optarg && strlen( optarg )? strdup( optarg ): NULL;
+                args->masterPasswordFD = optarg && strlen( optarg )? mpw_strdup( optarg ): NULL;
                 break;
             case 'M':
                 // Passing your master password via the command-line is insecure.  Testing purposes only.
-                args->masterPassword = optarg && strlen( optarg )? strdup( optarg ): NULL;
+                args->masterPassword = optarg && strlen( optarg )? mpw_strdup( optarg ): NULL;
                 break;
             case 't':
-                args->resultType = optarg && strlen( optarg )? strdup( optarg ): NULL;
+                args->resultType = optarg && strlen( optarg )? mpw_strdup( optarg ): NULL;
                 break;
             case 'P':
-                args->resultParam = optarg && strlen( optarg )? strdup( optarg ): NULL;
+                args->resultParam = optarg && strlen( optarg )? mpw_strdup( optarg ): NULL;
                 break;
             case 'c':
-                args->siteCounter = optarg && strlen( optarg )? strdup( optarg ): NULL;
+                args->siteCounter = optarg && strlen( optarg )? mpw_strdup( optarg ): NULL;
                 break;
             case 'a':
-                args->algorithmVersion = optarg && strlen( optarg )? strdup( optarg ): NULL;
+                args->algorithmVersion = optarg && strlen( optarg )? mpw_strdup( optarg ): NULL;
                 break;
             case 'p':
-                args->keyPurpose = optarg && strlen( optarg )? strdup( optarg ): NULL;
+                args->keyPurpose = optarg && strlen( optarg )? mpw_strdup( optarg ): NULL;
                 break;
             case 'C':
-                args->keyContext = optarg && strlen( optarg )? strdup( optarg ): NULL;
+                args->keyContext = optarg && strlen( optarg )? mpw_strdup( optarg ): NULL;
                 break;
             case 'f':
-                args->sitesFormat = optarg && strlen( optarg )? strdup( optarg ): NULL;
+                args->sitesFormat = optarg && strlen( optarg )? mpw_strdup( optarg ): NULL;
                 operation->sitesFormatFixed = false;
                 break;
             case 'F':
-                args->sitesFormat = optarg && strlen( optarg )? strdup( optarg ): NULL;
+                args->sitesFormat = optarg && strlen( optarg )? mpw_strdup( optarg ): NULL;
                 operation->sitesFormatFixed = true;
                 break;
             case 'R':
-                args->sitesRedacted = optarg && strlen( optarg )? strdup( optarg ): NULL;
+                args->sitesRedacted = optarg && strlen( optarg )? mpw_strdup( optarg ): NULL;
                 break;
             case 'v':
                 ++mpw_verbosity;
@@ -330,31 +335,33 @@ void cli_args(Arguments *args, Operation *operation, const int argc, char *const
             case '?':
                 switch (optopt) {
                     case 'u':
-                        ftl( "Missing full name to option: -%c\n", optopt );
+                        ftl( "Missing full name to option: -%c", optopt );
                         exit( EX_USAGE );
                     case 't':
-                        ftl( "Missing type name to option: -%c\n", optopt );
+                        ftl( "Missing type name to option: -%c", optopt );
                         exit( EX_USAGE );
                     case 'c':
-                        ftl( "Missing counter value to option: -%c\n", optopt );
+                        ftl( "Missing counter value to option: -%c", optopt );
                         exit( EX_USAGE );
                     default:
-                        ftl( "Unknown option: -%c\n", optopt );
+                        ftl( "Unknown option: -%c", optopt );
                         exit( EX_USAGE );
                 }
             default:
-                ftl( "Unexpected option: %c\n", opt );
+                ftl( "Unexpected option: %c", opt );
                 exit( EX_USAGE );
         }
 
     if (optind < argc && argv[optind])
-        args->siteName = strdup( argv[optind] );
+        args->siteName = mpw_strdup( argv[optind] );
 }
 
 void cli_fullName(Arguments *args, Operation *operation) {
 
-    if ((!operation->fullName || !strlen( operation->fullName )) && args->fullName)
-        operation->fullName = strdup( args->fullName );
+    mpw_free_string( &operation->fullName );
+
+    if (args->fullName)
+        operation->fullName = mpw_strdup( args->fullName );
 
     if (!operation->fullName || !strlen( operation->fullName ))
         do {
@@ -362,7 +369,7 @@ void cli_fullName(Arguments *args, Operation *operation) {
         } while (operation->fullName && !strlen( operation->fullName ));
 
     if (!operation->fullName || !strlen( operation->fullName )) {
-        ftl( "Missing full name.\n" );
+        ftl( "Missing full name." );
         cli_free( args, operation );
         exit( EX_DATAERR );
     }
@@ -370,14 +377,16 @@ void cli_fullName(Arguments *args, Operation *operation) {
 
 void cli_masterPassword(Arguments *args, Operation *operation) {
 
-    if ((!operation->masterPassword || !strlen( operation->masterPassword )) && args->masterPasswordFD) {
+    mpw_free_string( &operation->masterPassword );
+
+    if (args->masterPasswordFD) {
         operation->masterPassword = mpw_read_fd( atoi( args->masterPasswordFD ) );
         if (!operation->masterPassword && errno)
-            wrn( "Error reading master password from FD %s: %s\n", args->masterPasswordFD, strerror( errno ) );
+            wrn( "Error reading master password from FD %s: %s", args->masterPasswordFD, strerror( errno ) );
     }
 
-    if ((!operation->masterPassword || !strlen( operation->masterPassword )) && args->masterPassword)
-        operation->masterPassword = strdup( args->masterPassword );
+    if (args->masterPassword && !operation->masterPassword)
+        operation->masterPassword = mpw_strdup( args->masterPassword );
 
     if (!operation->masterPassword || !strlen( operation->masterPassword ))
         do {
@@ -385,7 +394,7 @@ void cli_masterPassword(Arguments *args, Operation *operation) {
         } while (operation->masterPassword && !strlen( operation->masterPassword ));
 
     if (!operation->masterPassword || !strlen( operation->masterPassword )) {
-        ftl( "Missing master password.\n" );
+        ftl( "Missing master password." );
         cli_free( args, operation );
         exit( EX_DATAERR );
     }
@@ -393,15 +402,15 @@ void cli_masterPassword(Arguments *args, Operation *operation) {
 
 void cli_siteName(Arguments *args, Operation *operation) {
 
-    if ((!operation->siteName || !strlen( operation->siteName )) && args->siteName)
-        operation->siteName = strdup( args->siteName );
-    if (!operation->siteName || !strlen( operation->siteName ))
-        do {
-            operation->siteName = mpw_getline( "Site name:" );
-        } while (operation->siteName && !strlen( operation->siteName ));
+    mpw_free_string( &operation->siteName );
 
-    if (!operation->siteName || !strlen( operation->siteName )) {
-        ftl( "Missing site name.\n" );
+    if (args->siteName)
+        operation->siteName = mpw_strdup( args->siteName );
+    if (!operation->siteName)
+        operation->siteName = mpw_getline( "Site name:" );
+
+    if (!operation->siteName) {
+        ftl( "Missing site name." );
         cli_free( args, operation );
         exit( EX_DATAERR );
     }
@@ -414,7 +423,7 @@ void cli_sitesFormat(Arguments *args, Operation *operation) {
 
     operation->sitesFormat = mpw_formatWithName( args->sitesFormat );
     if (ERR == (int)operation->sitesFormat) {
-        ftl( "Invalid sites format: %s\n", args->sitesFormat );
+        ftl( "Invalid sites format: %s", args->sitesFormat );
         cli_free( args, operation );
         exit( EX_DATAERR );
     }
@@ -427,7 +436,7 @@ void cli_keyPurpose(Arguments *args, Operation *operation) {
 
     operation->keyPurpose = mpw_purposeWithName( args->keyPurpose );
     if (ERR == (int)operation->keyPurpose) {
-        ftl( "Invalid purpose: %s\n", args->keyPurpose );
+        ftl( "Invalid purpose: %s", args->keyPurpose );
         cli_free( args, operation );
         exit( EX_DATAERR );
     }
@@ -438,7 +447,7 @@ void cli_keyContext(Arguments *args, Operation *operation) {
     if (!args->keyContext)
         return;
 
-    operation->keyContext = strdup( args->keyContext );
+    operation->keyContext = mpw_strdup( args->keyContext );
 }
 
 void cli_user(Arguments *args, Operation *operation) {
@@ -446,18 +455,18 @@ void cli_user(Arguments *args, Operation *operation) {
     // Find mpsites file from parameters.
     FILE *sitesFile = NULL;
     mpw_free_string( &operation->sitesPath );
-    operation->sitesPath = mpw_path( operation->fullName, mpw_marshall_format_extension( operation->sitesFormat ) );
+    operation->sitesPath = mpw_path( operation->fullName, mpw_marshal_format_extension( operation->sitesFormat ) );
     if (!operation->sitesPath || !(sitesFile = fopen( operation->sitesPath, "r" ))) {
-        dbg( "Couldn't open configuration file:\n  %s: %s\n", operation->sitesPath, strerror( errno ) );
+        dbg( "Couldn't open configuration file:\n  %s: %s", operation->sitesPath, strerror( errno ) );
 
         // Try to fall back to the flat format.
         if (!operation->sitesFormatFixed) {
             mpw_free_string( &operation->sitesPath );
-            operation->sitesPath = mpw_path( operation->fullName, mpw_marshall_format_extension( MPMarshallFormatFlat ) );
+            operation->sitesPath = mpw_path( operation->fullName, mpw_marshal_format_extension( MPMarshalFormatFlat ) );
             if (operation->sitesPath && (sitesFile = fopen( operation->sitesPath, "r" )))
-                operation->sitesFormat = MPMarshallFormatFlat;
+                operation->sitesFormat = MPMarshalFormatFlat;
             else
-                dbg( "Couldn't open configuration file:\n  %s: %s\n", operation->sitesPath, strerror( errno ) );
+                dbg( "Couldn't open configuration file:\n  %s: %s", operation->sitesPath, strerror( errno ) );
         }
     }
 
@@ -469,46 +478,47 @@ void cli_user(Arguments *args, Operation *operation) {
         // Read file.
         char *sitesInputData = mpw_read_file( sitesFile );
         if (ferror( sitesFile ))
-            wrn( "Error while reading configuration file:\n  %s: %d\n", operation->sitesPath, ferror( sitesFile ) );
+            wrn( "Error while reading configuration file:\n  %s: %d", operation->sitesPath, ferror( sitesFile ) );
         fclose( sitesFile );
 
         // Parse file.
-        MPMarshallInfo *sitesInputInfo = mpw_marshall_read_info( sitesInputData );
-        MPMarshallFormat sitesInputFormat = args->sitesFormat? operation->sitesFormat: sitesInputInfo->format;
-        MPMarshallError marshallError = { .type = MPMarshallSuccess };
+        MPMarshalInfo *sitesInputInfo = mpw_marshal_read_info( sitesInputData );
+        MPMarshalFormat sitesInputFormat = args->sitesFormat? operation->sitesFormat: sitesInputInfo->format;
+        MPMarshalError marshalError = { .type = MPMarshalSuccess };
         mpw_marshal_info_free( &sitesInputInfo );
-        operation->user = mpw_marshall_read( sitesInputData, sitesInputFormat, operation->masterPassword, &marshallError );
-        if (marshallError.type == MPMarshallErrorMasterPassword && operation->allowPasswordUpdate) {
+        mpw_marshal_free( &operation->user );
+        operation->user = mpw_marshal_read( sitesInputData, sitesInputFormat, operation->masterPassword, &marshalError );
+        if (marshalError.type == MPMarshalErrorMasterPassword && operation->allowPasswordUpdate) {
             // Update master password in mpsites.
-            while (marshallError.type == MPMarshallErrorMasterPassword) {
-                inf( "Given master password does not match configuration.\n" );
-                inf( "To update the configuration with this new master password, first confirm the old master password.\n" );
+            while (marshalError.type == MPMarshalErrorMasterPassword) {
+                inf( "Given master password does not match configuration." );
+                inf( "To update the configuration with this new master password, first confirm the old master password." );
 
                 const char *importMasterPassword = NULL;
                 while (!importMasterPassword || !strlen( importMasterPassword ))
                     importMasterPassword = mpw_getpass( "Old master password: " );
 
                 mpw_marshal_free( &operation->user );
-                operation->user = mpw_marshall_read( sitesInputData, sitesInputFormat, importMasterPassword, &marshallError );
+                operation->user = mpw_marshal_read( sitesInputData, sitesInputFormat, importMasterPassword, &marshalError );
                 mpw_free_string( &importMasterPassword );
             }
             if (operation->user) {
                 mpw_free_string( &operation->user->masterPassword );
-                operation->user->masterPassword = strdup( operation->masterPassword );
+                operation->user->masterPassword = mpw_strdup( operation->masterPassword );
             }
         }
         mpw_free_string( &sitesInputData );
 
         // Incorrect master password.
-        if (marshallError.type == MPMarshallErrorMasterPassword) {
-            ftl( "Incorrect master password according to configuration:\n  %s: %s\n", operation->sitesPath, marshallError.description );
+        if (marshalError.type == MPMarshalErrorMasterPassword) {
+            ftl( "Incorrect master password according to configuration:\n  %s: %s", operation->sitesPath, marshalError.description );
             cli_free( args, operation );
             exit( EX_DATAERR );
         }
 
         // Any other parse error.
-        if (!operation->user || marshallError.type != MPMarshallSuccess) {
-            err( "Couldn't parse configuration file:\n  %s: %s\n", operation->sitesPath, marshallError.description );
+        if (!operation->user || marshalError.type != MPMarshalSuccess) {
+            err( "Couldn't parse configuration file:\n  %s: %s", operation->sitesPath, marshalError.description );
             cli_free( args, operation );
             exit( EX_DATAERR );
         }
@@ -516,7 +526,7 @@ void cli_user(Arguments *args, Operation *operation) {
 
     // If no user from mpsites, create a new one.
     if (!operation->user)
-        operation->user = mpw_marshall_user(
+        operation->user = mpw_marshal_user(
                 operation->fullName, operation->masterPassword, MPAlgorithmVersionCurrent );
 }
 
@@ -532,8 +542,8 @@ void cli_site(Arguments __unused *args, Operation *operation) {
 
     // If no site from mpsites, create a new one.
     if (!operation->site)
-        operation->site = mpw_marshall_site(
-                operation->user, operation->siteName, MPResultTypeDefault, MPCounterValueDefault, operation->user->algorithm );
+        operation->site = mpw_marshal_site(
+                operation->user, operation->siteName, operation->user->defaultType, MPCounterValueDefault, operation->user->algorithm );
 }
 
 void cli_question(Arguments __unused *args, Operation *operation) {
@@ -549,7 +559,7 @@ void cli_question(Arguments __unused *args, Operation *operation) {
         case MPKeyPurposeRecovery:
             for (size_t q = 0; !operation->question && q < operation->site->questions_count; ++q)
                 if ((!operation->keyContext && !strlen( (&operation->site->questions[q])->keyword )) ||
-                    (operation->keyContext && strcmp( (&operation->site->questions[q])->keyword, operation->keyContext ) != 0))
+                    (operation->keyContext && strcmp( (&operation->site->questions[q])->keyword, operation->keyContext ) == 0))
                     operation->question = &operation->site->questions[q];
 
             // If no question from mpsites, create a new one.
@@ -561,7 +571,8 @@ void cli_question(Arguments __unused *args, Operation *operation) {
 
 void cli_operation(Arguments __unused *args, Operation *operation) {
 
-    operation->identicon = mpw_identicon( operation->user->fullName, operation->user->masterPassword );
+    mpw_free_string( &operation->identicon );
+    operation->identicon = mpw_identicon_str( mpw_identicon( operation->user->fullName, operation->user->masterPassword ) );
 
     if (!operation->site)
         abort();
@@ -570,23 +581,23 @@ void cli_operation(Arguments __unused *args, Operation *operation) {
         case MPKeyPurposeAuthentication: {
             operation->purposeResult = "password";
             operation->resultType = operation->site->type;
-            operation->resultState = operation->site->content? strdup( operation->site->content ): NULL;
+            operation->resultState = operation->site->content? mpw_strdup( operation->site->content ): NULL;
             operation->siteCounter = operation->site->counter;
             break;
         }
         case MPKeyPurposeIdentification: {
             operation->purposeResult = "login";
             operation->resultType = operation->site->loginType;
-            operation->resultState = operation->site->loginContent? strdup( operation->site->loginContent ): NULL;
+            operation->resultState = operation->site->loginContent? mpw_strdup( operation->site->loginContent ): NULL;
             operation->siteCounter = MPCounterValueInitial;
             break;
         }
         case MPKeyPurposeRecovery: {
             mpw_free_string( &operation->keyContext );
             operation->purposeResult = "answer";
-            operation->keyContext = operation->question->keyword? strdup( operation->question->keyword ): NULL;
+            operation->keyContext = operation->question->keyword? mpw_strdup( operation->question->keyword ): NULL;
             operation->resultType = operation->question->type;
-            operation->resultState = operation->question->content? strdup( operation->question->content ): NULL;
+            operation->resultState = operation->question->content? mpw_strdup( operation->question->content ): NULL;
             operation->siteCounter = MPCounterValueInitial;
             break;
         }
@@ -602,7 +613,7 @@ void cli_resultType(Arguments *args, Operation *operation) {
 
     operation->resultType = mpw_typeWithName( args->resultType );
     if (ERR == (int)operation->resultType) {
-        ftl( "Invalid type: %s\n", args->resultType );
+        ftl( "Invalid type: %s", args->resultType );
         cli_free( args, operation );
         exit( EX_USAGE );
     }
@@ -631,7 +642,7 @@ void cli_siteCounter(Arguments *args, Operation *operation) {
 
     long long int siteCounterInt = atoll( args->siteCounter );
     if (siteCounterInt < MPCounterValueFirst || siteCounterInt > MPCounterValueLast) {
-        ftl( "Invalid site counter: %s\n", args->siteCounter );
+        ftl( "Invalid site counter: %s", args->siteCounter );
         cli_free( args, operation );
         exit( EX_USAGE );
     }
@@ -652,7 +663,8 @@ void cli_resultParam(Arguments *args, Operation *operation) {
     if (!args->resultParam)
         return;
 
-    operation->resultParam = strdup( args->resultParam );
+    mpw_free_string( &operation->resultParam );
+    operation->resultParam = mpw_strdup( args->resultParam );
 }
 
 void cli_algorithmVersion(Arguments *args, Operation *operation) {
@@ -664,7 +676,7 @@ void cli_algorithmVersion(Arguments *args, Operation *operation) {
 
     int algorithmVersionInt = atoi( args->algorithmVersion );
     if (algorithmVersionInt < MPAlgorithmVersionFirst || algorithmVersionInt > MPAlgorithmVersionLast) {
-        ftl( "Invalid algorithm version: %s\n", args->algorithmVersion );
+        ftl( "Invalid algorithm version: %s", args->algorithmVersion );
         cli_free( args, operation );
         exit( EX_USAGE );
     }
@@ -677,7 +689,7 @@ void cli_sitesRedacted(Arguments *args, Operation *operation) {
         operation->user->redacted = strcmp( args->sitesRedacted, "1" ) == 0;
 
     else if (!operation->user->redacted)
-        wrn( "Sites configuration is not redacted.  Use -R 1 to change this.\n" );
+        wrn( "Sites configuration is not redacted.  Use -R 1 to change this." );
 }
 
 void cli_mpw(Arguments *args, Operation *operation) {
@@ -685,13 +697,15 @@ void cli_mpw(Arguments *args, Operation *operation) {
     if (!operation->site)
         abort();
 
-    inf( "%s's %s for %s:\n[ %s ]: ", operation->user->fullName, operation->purposeResult, operation->site->name, operation->identicon );
+    if (mpw_verbosity >= inf_level)
+        fprintf( stderr, "%s's %s for %s:\n[ %s ]: ",
+                operation->user->fullName, operation->purposeResult, operation->site->name, operation->identicon );
 
     // Determine master key.
     MPMasterKey masterKey = mpw_masterKey(
             operation->user->fullName, operation->user->masterPassword, operation->site->algorithm );
     if (!masterKey) {
-        ftl( "Couldn't derive master key.\n" );
+        ftl( "Couldn't derive master key." );
         cli_free( args, operation );
         exit( EX_SOFTWARE );
     }
@@ -702,7 +716,7 @@ void cli_mpw(Arguments *args, Operation *operation) {
         if (!(operation->resultState = mpw_siteState( masterKey, operation->site->name, operation->siteCounter,
                 operation->keyPurpose, operation->keyContext, operation->resultType, operation->resultParam,
                 operation->site->algorithm ))) {
-            ftl( "Couldn't encrypt site result.\n" );
+            ftl( "Couldn't encrypt site result." );
             mpw_free( &masterKey, MPMasterKeySize );
             cli_free( args, operation );
             exit( EX_SOFTWARE );
@@ -712,18 +726,18 @@ void cli_mpw(Arguments *args, Operation *operation) {
         switch (operation->keyPurpose) {
             case MPKeyPurposeAuthentication: {
                 mpw_free_string( &operation->site->content );
-                operation->site->content = strdup( operation->resultState );
+                operation->site->content = mpw_strdup( operation->resultState );
                 break;
             }
             case MPKeyPurposeIdentification: {
                 mpw_free_string( &operation->site->loginContent );
-                operation->site->loginContent = strdup( operation->resultState );
+                operation->site->loginContent = mpw_strdup( operation->resultState );
                 break;
             }
 
             case MPKeyPurposeRecovery: {
                 mpw_free_string( &operation->question->content );
-                operation->question->content = strdup( operation->resultState );
+                operation->question->content = mpw_strdup( operation->resultState );
                 break;
             }
         }
@@ -734,20 +748,21 @@ void cli_mpw(Arguments *args, Operation *operation) {
 
     // resultParam defaults to state.
     if (!operation->resultParam && operation->resultState)
-        operation->resultParam = strdup( operation->resultState );
+        operation->resultParam = mpw_strdup( operation->resultState );
 
     // Generate result.
     const char *result = mpw_siteResult( masterKey, operation->site->name, operation->siteCounter,
             operation->keyPurpose, operation->keyContext, operation->resultType, operation->resultParam, operation->site->algorithm );
     mpw_free( &masterKey, MPMasterKeySize );
     if (!result) {
-        ftl( "Couldn't generate site result.\n" );
+        ftl( "Couldn't generate site result." );
         cli_free( args, operation );
         exit( EX_SOFTWARE );
     }
+    fflush( NULL );
     fprintf( stdout, "%s\n", result );
     if (operation->site->url)
-        inf( "See: %s\n", operation->site->url );
+        inf( "See: %s", operation->site->url );
     mpw_free_string( &result );
 
     // Update usage metadata.
@@ -757,27 +772,29 @@ void cli_mpw(Arguments *args, Operation *operation) {
 
 void cli_save(Arguments __unused *args, Operation *operation) {
 
-    if (operation->sitesFormat == MPMarshallFormatNone)
+    if (operation->sitesFormat == MPMarshalFormatNone)
         return;
 
     if (!operation->sitesFormatFixed)
-        operation->sitesFormat = MPMarshallFormatDefault;
-    operation->sitesPath = mpw_path( operation->user->fullName, mpw_marshall_format_extension( operation->sitesFormat ) );
-    dbg( "Updating: %s (%s)\n", operation->sitesPath, mpw_nameForFormat( operation->sitesFormat ) );
+        operation->sitesFormat = MPMarshalFormatDefault;
+
+    mpw_free_string( &operation->sitesPath );
+    operation->sitesPath = mpw_path( operation->user->fullName, mpw_marshal_format_extension( operation->sitesFormat ) );
+    dbg( "Updating: %s (%s)", operation->sitesPath, mpw_nameForFormat( operation->sitesFormat ) );
 
     FILE *sitesFile = NULL;
     if (!operation->sitesPath || !mpw_mkdirs( operation->sitesPath ) || !(sitesFile = fopen( operation->sitesPath, "w" ))) {
-        wrn( "Couldn't create updated configuration file:\n  %s: %s\n", operation->sitesPath, strerror( errno ) );
+        wrn( "Couldn't create updated configuration file:\n  %s: %s", operation->sitesPath, strerror( errno ) );
         return;
     }
 
     char *buf = NULL;
-    MPMarshallError marshallError = { .type = MPMarshallSuccess };
-    if (!mpw_marshall_write( &buf, operation->sitesFormat, operation->user, &marshallError ) || marshallError.type != MPMarshallSuccess)
-        wrn( "Couldn't encode updated configuration file:\n  %s: %s\n", operation->sitesPath, marshallError.description );
+    MPMarshalError marshalError = { .type = MPMarshalSuccess };
+    if (!mpw_marshal_write( &buf, operation->sitesFormat, operation->user, &marshalError ) || marshalError.type != MPMarshalSuccess)
+        wrn( "Couldn't encode updated configuration file:\n  %s: %s", operation->sitesPath, marshalError.description );
 
     else if (fwrite( buf, sizeof( char ), strlen( buf ), sitesFile ) != strlen( buf ))
-        wrn( "Error while writing updated configuration file:\n  %s: %d\n", operation->sitesPath, ferror( sitesFile ) );
+        wrn( "Error while writing updated configuration file:\n  %s: %d", operation->sitesPath, ferror( sitesFile ) );
 
     mpw_free_string( &buf );
     fclose( sitesFile );
