@@ -1,10 +1,27 @@
+//==============================================================================
+// This file is part of Master Password.
+// Copyright (c) 2011-2017, Maarten Billemont.
+//
+// Master Password is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Master Password is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You can find a copy of the GNU General Public License in the
+// LICENSE file.  Alternatively, see <http://www.gnu.org/licenses/>.
+//==============================================================================
+
 package com.lyndir.masterpassword.gui.view;
 
 import com.google.common.collect.ImmutableList;
 import com.lyndir.masterpassword.gui.Res;
-import com.lyndir.masterpassword.gui.model.User;
 import com.lyndir.masterpassword.gui.util.Components;
-import com.lyndir.masterpassword.gui.view.UnlockFrame;
+import com.lyndir.masterpassword.model.MPUser;
 import java.awt.*;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -14,7 +31,7 @@ import javax.swing.*;
 /**
  * @author lhunath, 2014-06-11
  */
-public abstract class AuthenticationPanel extends Components.GradientPanel {
+public abstract class AuthenticationPanel<U extends MPUser<?>> extends Components.GradientPanel {
 
     protected final UnlockFrame unlockFrame;
     protected final JLabel      avatarLabel;
@@ -47,11 +64,12 @@ public abstract class AuthenticationPanel extends Components.GradientPanel {
     }
 
     @Nullable
-    protected abstract User getSelectedUser();
+    protected abstract U getSelectedUser();
 
     @Nonnull
     public abstract char[] getMasterPassword();
 
+    @Nullable
     public Component getFocusComponent() {
         return null;
     }
@@ -61,4 +79,6 @@ public abstract class AuthenticationPanel extends Components.GradientPanel {
     }
 
     public abstract void reset();
+
+    public abstract PasswordFrame<?, ?> newPasswordFrame();
 }
